+++ ecopy/diversity/div_partition.py
@@ -1,55 +1,40 @@
 import numpy as np
-from pandas import DataFrame 
+from pandas import DataFrame
 
-def diversity(x, method='shannon', breakNA=True, num_equiv=True):
+def div_partition(x, method='shannon', breakNA=True, weights=None):
 	'''
 	Docstring for function ecopy.diversity
 	========================
-	Computes a given diversity index for a site x species matrix.
-		All indices computed along rows (axis = 1)
+	Decomposes diversity measures into alpha, beta, and gamma components.
 
 	Use
 	----
-	diversity(x, method='shannon', breakNA=True)
+	div_partition(x, method='shannon', breakNA=True)
 
-	Returns a numpy.ndarray or a pandas.Series
+	Returns a tuple of alpha, beta, and gamma diversity
 
 	Parameters
 	----------
 	x:  numpy array or pandas dataframe with observations as rows
 		and descriptors as columns
-	method: a method used for calculating species diversity
-		shannon: Shannon's H
-			H = -sum( p_i * log(p_i) )
-			where p_i is relative abundance of species i
-		gini-simpson: Gini-Simpson
-			D = 1 - sum( p_i^2 )
-		simpson: Simpson's D
-			D = sum(p_i^2)
-		dominance: Dominance index
-			D = max(p_i)
-		spRich: Species richness
-			D = Number of non-zero observations
-		even: Evenness of a sample. This is Shannon's H divided by the
-			log of species richness
+	method: a method used for calculating species diversity. See the 
+		diversity() function for details. Can only be 'shannon', 
+		'gini-simpson', 'simpson', or 'spRich'.
 	breakNA: should the process halt if the matrix contains any NAs?
 		if False, then NA's undergo pairwise deletion during distance calculation,
 		such that when calculating the distance between two rows, if any
 		species is missing from a row, that species is removed from both rows
-	num_equiv: Whether or not the diversity index return species number equivalents,
-		i.e. the number of species of identical abundance. This has better properties 
-		than raw diversity. The number equivalents are as follows:
-
-		shannon: np.exp(H)
-		gini-simpson: 1 / (1-D)
-		simpson: 1/D
+	weights: weights given for each row (site). Defaults to the 
+		sum of each row divided by the sum of the matrix. This yields
+		weights based on the number of individuals in a site for raw
+		abundance data or equal weights for relative abundance data.
 
 	Example
 	--------
 	import ecopy as ep
 	varespec = ep.load_data('varespec')
 	
-	div = ep.diversity(varespec, 'shannon')
+	D_alpha, D_beta, D_gamma = ep.div_partition(varespec, 'shannon')
 	'''
 	listofmethods = ['shannon', 'gini-simpson', 'simpson', 'dominance', 'spRich', 'even']
 	if not isinstance(breakNA, bool):
@@ -81,31 +66,44 @@ def diversity(x, method='shannon', breakNA=True, num_equiv=True):
 			msg = 'Array contains negative values'
 			raise ValueError(msg)
 		z = np.array(x, 'float')
-	z = z / z.sum(axis=1)[:,np.newaxis]
+	if weights is None:
+		weights = z.sum(axis=1)/z.sum()
+	w = np.array(weights)/np.array(weights).sum()
+	relMat = z / z.sum(axis=1)[:,np.newaxis]
+	totalList = z.sum(axis=0)
+	totalRel = totalList / totalList.sum()
 	if method=='shannon':
-		div = np.apply_along_axis(shannonFunc, 1, z)
-		if num_equiv:
-			div = np.exp(div)
-		return div
+		div = np.apply_along_axis(shannonFunc, 1, relMat)
+		H_alpha = (w*div).sum() / w.sum()
+		H_gamma = shannonFunc(totalRel)
+		D_alpha = np.exp(H_alpha)
+		D_gamma = np.exp(H_gamma)
+		D_beta = D_gamma / D_alpha
+		return D_alpha, D_beta, D_gamma
 	if method=='gini-simpson':
-		div = np.apply_along_axis(giniFunc, 1, z)
-		if num_equiv:
-			div = 1.(1.-div)
-		return div
+		div = np.apply_along_axis(giniFunc, 1, relMat)
+		H_alpha = (w*div).sum()/w.sum()
+		H_gamma = giniFunc(totalRel)
+		D_alpha = 1./(1.-H_alpha)
+		D_gamma = 1./(1.-H_gamma)
+		D_beta = D_gamma / D_alpha
+		return D_alpha, D_beta, D_gamma
 	if method=='simpson':
-		div = np.apply_along_axis(simpson, 1, z)
-		if num_equiv:
-			div = 1./divpe
-		return div
-	if method=='dominance':
-		div = np.apply_along_axis(dom, 1, z)
-		return div
+		div = np.apply_along_axis(simpson, 1, relMat)
+		H_alpha = (w*div).sum()/w.sum()
+		H_gamma = simpson(totalRel)
+		D_alpha = 1./H_alpha
+		D_gamma = 1./H_gamma
+		D_beta = D_gamma / D_alpha
+		return D_alpha, D_beta, D_gamma
 	if method=='spRich':
-		div = np.apply_along_axis(richness, 1, z)
-		return div
-	if method=='even':
-		div = np.apply_along_axis(evenFunc, 1, z)
-		return div
+		div = np.apply_along_axis(richness, 1, relMat)
+		H_alpha = (w*div).sum()/w.sum()
+		H_gamma = richness(totalRel)
+		D_alpha = H_alpha
+		D_gamma = H_gamma
+		D_beta = D_gamma / D_alpha
+		return D_alpha, D_beta, D_gamma
 
 def shannonFunc(y):
 	notabs = ~np.isnan(y)
@@ -126,22 +124,8 @@ def simpson(y):
 	D = np.sum( t**2 )
 	return D
 
-def dom(y):
-	notabs = ~np.isnan(y)
-	t = y[notabs] / np.sum(y[notabs])
-	D = np.max(t)
-	return D
-
 def richness(y):
 	notabs = ~np.isnan(y)
 	t = y[notabs]
 	D = np.sum(t!=0)
 	return float(D)
-
-def evenFunc(y):
-	notabs = ~np.isnan(y)
-	t = y[notabs] / np.sum(y[notabs])
-	n = float(np.sum(t!=0))
-	t = t[t!=0]
-	H = -np.sum( t*np.log(t) )
-	return H/np.log(n)